Yet another intthat remained

+++ libs/huge_integers.h
@@ -1,4 +1,4 @@
-typedef int* huge_int;
+typedef long* huge_int;
 
 void copy_huge_int(long s, huge_int a, huge_int b); //Takes the integer in a and puts it in b
 void huge_sum(long s, huge_int a, huge_int b, huge_int r);
